Create app.js as entry point to app.
