+++ bin/www
@@ -1,6 +1,6 @@
 require('dotenv').config();
 const http = require('http');
-const app = require('../server');
+const app = require('../app');
 
 const port = parseInt(process.env.DB_PORT, 10) || 3000;
 app.set('port', port);
