+++ package.json
@@ -2,7 +2,7 @@
   "name": "couch-potatoes",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
     "build": "webpack --env.dev --env.client --env.server",
     "build:server": "webpack --env.server",
